+++ src/main/java/com/jagex/runescape/Class17.java
@@ -23,17 +23,17 @@ public class Class17 {
     public static int anInt464 = 0;
     public static int[] anIntArray466 = new int[256];
 
-    public static void method273(Class64 arg0, int arg1, boolean arg2) {
-        if(Class57.aClass64_1345 != null) {
+    public static void method273(GameSocket arg0, boolean arg2) {
+        if(Class57.gameSocket != null) {
             try {
-                Class57.aClass64_1345.method1009();
+                Class57.gameSocket.method1009();
             } catch(Exception exception) {
                 exception.printStackTrace();
                 /* empty */
             }
-            Class57.aClass64_1345 = null;
+            Class57.gameSocket = null;
         }
-        Class57.aClass64_1345 = arg0;
+        Class57.gameSocket = arg0;
         GameShell.method19(arg2, 24041);
         LinkedList.aClass40_Sub1_1081.currentPosition = 0;
         Class40_Sub5_Sub13.aClass40_Sub1_2752 = null;
@@ -62,21 +62,19 @@ public class Class17 {
                 class40_sub1.putByte(4);
                 class40_sub1.putByte(Class8.aByte302);
                 class40_sub1.putShortBE(0);
-                Class57.aClass64_1345.method1010(4, (byte) -19, 0, class40_sub1.buffer);
+                Class57.gameSocket.method1010(4, 0, class40_sub1.buffer);
             } catch(java.io.IOException ioexception) {
                 ioexception.printStackTrace();
                 try {
-                    Class57.aClass64_1345.method1009();
+                    Class57.gameSocket.method1009();
                 } catch(Exception exception) {
                     exception.printStackTrace();
                     /* empty */
                 }
-                Class57.aClass64_1345 = null;
+                Class57.gameSocket = null;
                 MovedStatics.anInt2278++;
             }
         }
-        if(arg1 < 102)
-            method273(null, 92, true);
         Class22.anInt547 = 0;
         LinkedList.aLong1051 = System.currentTimeMillis();
     }
