+++ src/main/java/com/jagex/runescape/Class39.java
@@ -56,7 +56,7 @@ public class Class39 implements Runnable {
             MovedStatics.aClass22_189 = new Class22_Sub2_Sub1(); // Java Midi Based
             return true;
         } catch(Throwable throwable) {
-            Runnable_Impl1 runnable_impl1 = arg0.method391(1048576);
+            Runnable_Impl1 runnable_impl1 = arg0.method391();
             if(runnable_impl1 != null) {
                 MovedStatics.aClass22_189 = new Class22_Sub2_Sub2(arg0, runnable_impl1); // ???
                 return true;
