+++ src/main/java/com/jagex/runescape/Class48.java
@@ -319,9 +319,9 @@ public class Class48 {
     }
 
     public static void logout(int arg0) {
-        if(MovedStatics.gameConnection != null) {
-            MovedStatics.gameConnection.method1009();
-            MovedStatics.gameConnection = null;
+        if(MovedStatics.gameSocket != null) {
+            MovedStatics.gameSocket.method1009();
+            MovedStatics.gameSocket = null;
         }
         RSCanvas.clearCaches();
         Npc.currentScene.initToNull();
