+++ src/main/java/com/jagex/runescape/ProducingGraphicsBuffer.java
@@ -32,7 +32,7 @@ public abstract class ProducingGraphicsBuffer {
             try {
                 CollisionMap.aClass8_166 = new Class8_Sub1_Sub1(arg2, arg0);
             } catch (Throwable throwable_0_) {
-                if (arg1 <= Signlink.aString735.toLowerCase().indexOf("microsoft")) {
+                if (arg1 <= Signlink.javaVendor.toLowerCase().indexOf("microsoft")) {
                     try {
                         CollisionMap.aClass8_166 = new Class8_Sub2();
                         return;
