+++ src/main/java/com/jagex/runescape/cache/media/SpotAnimDefinition.java
@@ -52,9 +52,9 @@ public class SpotAnimDefinition extends CachedNode {
     public static void method552(boolean arg0) {
         try {
             if (Class40_Sub3.anInt2032 == 0) {
-                if (MovedStatics.gameConnection != null) {
-                    MovedStatics.gameConnection.method1009();
-                    MovedStatics.gameConnection = null;
+                if (MovedStatics.gameSocket != null) {
+                    MovedStatics.gameSocket.method1009();
+                    MovedStatics.gameSocket = null;
                 }
                 Class37.aBoolean871 = false;
                 Class40_Sub3.anInt2032 = 1;
@@ -69,7 +69,7 @@ public class SpotAnimDefinition extends CachedNode {
                     throw new IOException();
                 }
                 if (FloorDecoration.aSignlinkNode_607.anInt434 == 1) {
-                    MovedStatics.gameConnection = new Class64((Socket) FloorDecoration.aSignlinkNode_607.value, Main.signlink);
+                    MovedStatics.gameSocket = new GameSocket((Socket) FloorDecoration.aSignlinkNode_607.value, Main.signlink);
                     Class40_Sub3.anInt2032 = 2;
                     FloorDecoration.aSignlinkNode_607 = null;
                 }
@@ -84,12 +84,12 @@ public class SpotAnimDefinition extends CachedNode {
                 SceneCluster.packetBuffer.putByte(14);
                 int i = (int) (0x1fL & l >> 16);
                 SceneCluster.packetBuffer.putByte(i);
-                MovedStatics.gameConnection.method1010(2, (byte) -19, 0, SceneCluster.packetBuffer.buffer);
+                MovedStatics.gameSocket.method1010(2, 0, SceneCluster.packetBuffer.buffer);
                 Class40_Sub3.anInt2032 = 3;
                 IncomingPackets.incomingPacketBuffer.currentPosition = 0;
             }
             if (Class40_Sub3.anInt2032 == 3) {
-                int i = MovedStatics.gameConnection.read();
+                int i = MovedStatics.gameSocket.read();
                 if (i != 0) {
                     Main.displayMessageForResponseCode(i);
                     return;
@@ -100,12 +100,12 @@ public class SpotAnimDefinition extends CachedNode {
             if (Class40_Sub3.anInt2032 == 4) {
 
                 if (IncomingPackets.incomingPacketBuffer.currentPosition < 8) {
-                    int i = MovedStatics.gameConnection.method1014(-127);
+                    int i = MovedStatics.gameSocket.method1014(-127);
                     if (i > -IncomingPackets.incomingPacketBuffer.currentPosition + 8) {
                         i = -IncomingPackets.incomingPacketBuffer.currentPosition + 8;
                     }
                     if (i > 0) {
-                        MovedStatics.gameConnection.method1008(IncomingPackets.incomingPacketBuffer.currentPosition, i, -128, IncomingPackets.incomingPacketBuffer.buffer);
+                        MovedStatics.gameSocket.method1008(IncomingPackets.incomingPacketBuffer.currentPosition, i, -128, IncomingPackets.incomingPacketBuffer.buffer);
                         IncomingPackets.incomingPacketBuffer.currentPosition += i;
                     }
                 }
@@ -156,7 +156,7 @@ public class SpotAnimDefinition extends CachedNode {
                 MovedStatics.packetBuffer.putIntBE(CacheArchive.jingleCacheArchive.anInt216);
                 MovedStatics.packetBuffer.putIntBE(CacheArchive.clientScriptCacheArchive.anInt216);
                 MovedStatics.packetBuffer.putBytes(0, SceneCluster.packetBuffer.currentPosition, SceneCluster.packetBuffer.buffer);
-                MovedStatics.gameConnection.method1010(MovedStatics.packetBuffer.currentPosition, (byte) -19, 0, MovedStatics.packetBuffer.buffer);
+                MovedStatics.gameSocket.method1010(MovedStatics.packetBuffer.currentPosition, 0, MovedStatics.packetBuffer.buffer);
                 SceneCluster.packetBuffer.initOutCipher(seeds);
                 for (int i = 0; i < 4; i++) {
                     seeds[i] += 50;
@@ -167,8 +167,8 @@ public class SpotAnimDefinition extends CachedNode {
 
 
 
-            if (Class40_Sub3.anInt2032 == 6 && MovedStatics.gameConnection.method1014(-126) > 0) {
-                int i = MovedStatics.gameConnection.read();
+            if (Class40_Sub3.anInt2032 == 6 && MovedStatics.gameSocket.method1014(-126) > 0) {
+                int i = MovedStatics.gameSocket.read();
                 if (i != 21 || Class51.currentAction != 20) {
                     if (i == 2) {
                         Class40_Sub3.anInt2032 = 9;
@@ -189,8 +189,8 @@ public class SpotAnimDefinition extends CachedNode {
                     Class40_Sub3.anInt2032 = 7;
                 }
             }
-            if (Class40_Sub3.anInt2032 == 7 && MovedStatics.gameConnection.method1014(-128) > 0) {
-                Class33.anInt784 = 180 + MovedStatics.gameConnection.read() * 60;
+            if (Class40_Sub3.anInt2032 == 7 && MovedStatics.gameSocket.method1014(-128) > 0) {
+                Class33.anInt784 = 180 + MovedStatics.gameSocket.read() * 60;
                 Class40_Sub3.anInt2032 = 8;
 
             }
@@ -202,27 +202,27 @@ public class SpotAnimDefinition extends CachedNode {
                 }
             } else {
 
-                if (Class40_Sub3.anInt2032 == 9 && MovedStatics.gameConnection.method1014(-121) >= 8) {
+                if (Class40_Sub3.anInt2032 == 9 && MovedStatics.gameSocket.method1014(-121) >= 8) {
                     Configuration.USERNAME = Native.username.toString();
                     Configuration.PASSWORD = Native.password.toString();
-                    InteractiveObject.playerRights = MovedStatics.gameConnection.read();
-                    Class22.accountFlagged = MovedStatics.gameConnection.read() == 1;
-                    PlayerAppearance.anInt708 = MovedStatics.gameConnection.read();
+                    InteractiveObject.playerRights = MovedStatics.gameSocket.read();
+                    Class22.accountFlagged = MovedStatics.gameSocket.read() == 1;
+                    PlayerAppearance.anInt708 = MovedStatics.gameSocket.read();
                     PlayerAppearance.anInt708 <<= 8;
-                    PlayerAppearance.anInt708 += MovedStatics.gameConnection.read();
-                    Class44.anInt1049 = MovedStatics.gameConnection.read();
-                    MovedStatics.gameConnection.method1008(0, 1, -127, IncomingPackets.incomingPacketBuffer.buffer);
+                    PlayerAppearance.anInt708 += MovedStatics.gameSocket.read();
+                    Class44.anInt1049 = MovedStatics.gameSocket.read();
+                    MovedStatics.gameSocket.method1008(0, 1, -127, IncomingPackets.incomingPacketBuffer.buffer);
                     IncomingPackets.incomingPacketBuffer.currentPosition = 0;
                     IncomingPackets.incomingPacket = IncomingPackets.incomingPacketBuffer.getPacket();
-                    MovedStatics.gameConnection.method1008(0, 2, -127, IncomingPackets.incomingPacketBuffer.buffer);
+                    MovedStatics.gameSocket.method1008(0, 2, -127, IncomingPackets.incomingPacketBuffer.buffer);
                     IncomingPackets.incomingPacketBuffer.currentPosition = 0;
                     IncomingPackets.incomingPacketSize = IncomingPackets.incomingPacketBuffer.getUnsignedShortBE();
                     Class40_Sub3.anInt2032 = 10;
                 }
                 if (Class40_Sub3.anInt2032 == 10) {
-                    if (MovedStatics.gameConnection.method1014(-124) >= IncomingPackets.incomingPacketSize) {
+                    if (MovedStatics.gameSocket.method1014(-124) >= IncomingPackets.incomingPacketSize) {
                         IncomingPackets.incomingPacketBuffer.currentPosition = 0;
-                        MovedStatics.gameConnection.method1008(0, IncomingPackets.incomingPacketSize, -128, IncomingPackets.incomingPacketBuffer.buffer);
+                        MovedStatics.gameSocket.method1008(0, IncomingPackets.incomingPacketSize, -128, IncomingPackets.incomingPacketBuffer.buffer);
                         Main.method44();
                         Class51.regionX = -1;
                         FloorDecoration.method343(false, 5688);
@@ -303,7 +303,7 @@ public class SpotAnimDefinition extends CachedNode {
     public Model method549(int arg0, int arg1) {
         Model class40_sub5_sub17_sub5 = (Model) MovedStatics.aClass9_851.get((long) id);
         if (class40_sub5_sub17_sub5 == null) {
-            class40_sub5_sub17_sub5 = Model.getModel(UnderlayDefinition.aCacheArchive_2582, modelId);
+            class40_sub5_sub17_sub5 = Model.getModel(UnderlayDefinition. aCacheArchive_2582, modelId);
             if (class40_sub5_sub17_sub5 == null) {
                 return null;
             }
