+++ src/main/java/com/jagex/runescape/media/renderable/actor/PlayerAppearance.java
@@ -23,7 +23,7 @@ public class PlayerAppearance {
     public static int[] anIntArray712 = new int[50];
     public static ProducingGraphicsBuffer tabPieveLowerRight;
     public static int[] anIntArray715 = new int[50];
-    public static Class64 aClass64_717;
+    public static GameSocket gameSocket;
 
     public static int identityKitLength;
     public static int[][] playerColours = {{6798, 107, 10283, 16, 4797, 7744, 5799, 4634, 33697, 22433, 2983, 54193}, {8741, 12, 64030, 43162, 7735, 8404, 1701, 38430, 24094, 10153, 56621, 4783, 1341, 16578, 35003, 25239}, {25238, 8742, 12, 64030, 43162, 7735, 8404, 1701, 38430, 24094, 10153, 56621, 4783, 1341, 16578, 35003}, {4626, 11146, 6439, 12, 4758, 10270}, {4550, 4537, 5681, 5673, 5790, 6806, 8076, 4574}};
